Implement NativeImplGenerator
