+++ Benchmarks/Benchmarks.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
@@ -10,6 +10,7 @@
     <PackageReference Include="CsvHelper" Version="30.0.1" />
     <PackageReference Include="MessagePack" Version="2.5.129" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageReference Include="RecordParser" Version="2.1.0" />
     <PackageReference Include="Utf8Json" Version="1.3.7" />
   </ItemGroup>
